Finished implementing networked maze generation
Now it works! :-D Please note that if the client walks out and triggers the rising walls before the server's connected and successfully sent the maze-sync RPC, the client will have an ungenerated (completely boxed-in) maze. This won't be an issue in the final game because we'll have a menu, and the game won't actually start until all players are connected.