Fix mouse-look input for players 3 and 4 (before they were receiving input from player 1's controls)
It should be noted that at the moment, although players 1 and 2 run on the server and 3 and 4 on the client, they each receive input from controllers 1, 2, 3, and 4 respectively. This is essentially an implicit assumption that both the client and server are running on the same computer (as they usually are during testing). :-) Normally, you would only expect the client to have two controllers plugged in, and to use those for its two players (3 and 4). This will need to be fixed later, but for now it'll be OK. Fixing this is not immediately trivial, because I don't (yet) know of a good way for the MouseLook and FPSInput scripts to check whether networking is active (they can't simply check the public bool in NetworkManager, since they're in a different assembly, being "standard assets" vs. "regular assets".)