Client now receives input for players 3 and 4 from controllers 1 and 2
After my last commit where I said this wasn't too important to fix, I realized it actually is, because otherwise we'll need to have four controllers plugged in to test players 3 and 4 at all (regardless of whether the client and server are on the same machine).

Don't ask about the "Network OnOff Foobar Thing". (Okay, so there's an explanation in the comments.)