Fix issue where walking sometimes wouldn't work (most showed up on levels 2 & 3)
The "Network OnOff Foobar Object" came back to bite me...fortunately it has now been put to rest. We still need a good way to inform CharacterMotor.js of whether networking is active (it's the only script that can't "see" the LockdownGlobals class, due to compilation order issues). The "best" fix I can think of at the moment would be to implement Singleton and LockdownGlobals in JavaScript instead of C#, ensuring that everything has access to them; but for now, other things are more important. The immediate consequence is that in 3-player local mode, player 3 can't sprint (or more precisely, player 3's sprinting is controlled by player 1's button).